sql: move row sampling to new stats package
